refactor: extract song templates into module
