+++ src/components/songTemplates.ts
@@ -1,0 +1,682 @@
+import type { Section, TemplateSpec } from "./SongForm";
+
+export const SECTION_PRESETS: Record<string, Section[]> = {
+  "Intro(4)-A(8)-B(8)-Break(4)-A(8)-Outro(4)": [
+    { name: "Intro", bars: 4, chords: [] },
+    { name: "A", bars: 8, chords: [] },
+    { name: "B", bars: 8, chords: [] },
+    { name: "Break", bars: 4, chords: [] },
+    { name: "A", bars: 8, chords: [] },
+    { name: "Outro", bars: 4, chords: [] },
+  ],
+  "A(16)-B(16)-Outro(8)": [
+    { name: "A", bars: 16, chords: [] },
+    { name: "B", bars: 16, chords: [] },
+    { name: "Outro", bars: 8, chords: [] },
+  ],
+  "A(8)x4": [
+    { name: "A", bars: 8, chords: [] },
+    { name: "A", bars: 8, chords: [] },
+    { name: "A", bars: 8, chords: [] },
+    { name: "A", bars: 8, chords: [] },
+  ],
+  "Intro(4)-A(8)-B(8)-C(8)-D(8)-E(8)-F(8)-Outro(4)": [
+    { name: "Intro", bars: 4, chords: [] },
+    { name: "A", bars: 8, chords: [] },
+    { name: "B", bars: 8, chords: [] },
+    { name: "C", bars: 8, chords: [] },
+    { name: "D", bars: 8, chords: [] },
+    { name: "E", bars: 8, chords: [] },
+    { name: "F", bars: 8, chords: [] },
+    { name: "Outro", bars: 4, chords: [] },
+  ],
+  "Intro(4)-A(7)-B(5)-C(7)-D(5)-E(7)-F(5)-Outro(4)": [
+    { name: "Intro", bars: 4, chords: [] },
+    { name: "A", bars: 7, chords: [] },
+    { name: "B", bars: 5, chords: [] },
+    { name: "C", bars: 7, chords: [] },
+    { name: "D", bars: 5, chords: [] },
+    { name: "E", bars: 7, chords: [] },
+    { name: "F", bars: 5, chords: [] },
+    { name: "Outro", bars: 4, chords: [] },
+  ],
+};
+export const PRESET_TEMPLATES: Record<string, TemplateSpec> = {
+  "Classic Lofi": {
+    structure: [
+      { name: "Intro", bars: 4, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "B", bars: 8, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "B", bars: 8, chords: [] },
+      { name: "Outro", bars: 4, chords: [] },
+    ],
+    bpm: 80,
+    key: "Auto",
+    mood: ["calm", "cozy", "nostalgic"],
+    instruments: ["rhodes", "nylon guitar", "upright bass"],
+    ambience: ["rain"],
+    drumPattern: "laidback",
+    variety: 45,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.02,
+    dither: true,
+    bpmJitterPct: 5,
+  },
+  "Study Session": {
+    structure: [
+      { name: "Intro", bars: 2, chords: [] },
+      { name: "A", bars: 16, chords: [] },
+      { name: "B", bars: 16, chords: [] },
+      { name: "A", bars: 16, chords: [] },
+      { name: "C", bars: 8, chords: [] },
+      { name: "Outro", bars: 2, chords: [] },
+    ],
+    bpm: 75,
+    key: "C",
+    mood: ["calm", "cozy"],
+    instruments: ["electric piano", "upright bass", "pads"],
+    ambience: ["cafe"],
+    drumPattern: "boom_bap_A",
+    variety: 30,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.01,
+    dither: true,
+    bpmJitterPct: 3,
+  },
+  "Jazz Cafe": {
+    structure: [
+      { name: "Intro", bars: 8, chords: [] },
+      { name: "A", bars: 12, chords: [] },
+      { name: "B", bars: 12, chords: [] },
+      { name: "Solo", bars: 8, chords: [] },
+      { name: "A", bars: 12, chords: [] },
+      { name: "Outro", bars: 8, chords: [] },
+    ],
+    bpm: 90,
+    key: "F",
+    mood: ["cozy", "nostalgic"],
+    instruments: ["piano", "upright bass", "pads"],
+    ambience: ["cafe"],
+    drumPattern: "swing",
+    variety: 20,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.0,
+    dither: true,
+    bpmJitterPct: 2,
+  },
+  "Midnight Drive": {
+    structure: [
+      { name: "Intro", bars: 4, chords: [] },
+      { name: "Verse", bars: 16, chords: [] },
+      { name: "Chorus", bars: 8, chords: [] },
+      { name: "Verse", bars: 16, chords: [] },
+      { name: "Chorus", bars: 8, chords: [] },
+      { name: "Bridge", bars: 8, chords: [] },
+      { name: "Outro", bars: 4, chords: [] },
+    ],
+    bpm: 85,
+    key: "D",
+    mood: ["melancholy", "nostalgic"],
+    instruments: ["clean electric guitar", "upright bass", "pads"],
+    ambience: ["rain"],
+    drumPattern: "half_time",
+    variety: 50,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.2,
+    dither: true,
+    bpmJitterPct: 4,
+  },
+  "Rain & Coffee": {
+    structure: [
+      { name: "Ambient", bars: 4, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "B", bars: 12, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "Ambient", bars: 4, chords: [] },
+    ],
+    bpm: 72,
+    key: "G",
+    mood: ["calm", "nostalgic"],
+    instruments: ["rhodes", "piano", "pads"],
+    ambience: ["rain", "cafe"],
+    drumPattern: "laidback",
+    variety: 40,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.05,
+    dither: true,
+    bpmJitterPct: 6,
+  },
+  "Bossa Nova": {
+    structure: [
+      { name: "Intro", bars: 4, chords: [] },
+      { name: "A", bars: 16, chords: [] },
+      { name: "B", bars: 16, chords: [] },
+      { name: "A", bars: 16, chords: [] },
+      { name: "Outro", bars: 4, chords: [] },
+    ],
+    bpm: 120,
+    key: "Auto",
+    mood: ["calm", "cozy"],
+    instruments: ["nylon guitar", "upright bass", "piano"],
+    ambience: ["cafe"],
+    drumPattern: "bossa_nova",
+    variety: 35,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.05,
+    dither: true,
+    bpmJitterPct: 3,
+  },
+  "Quick Beat": {
+    structure: [
+      { name: "Intro", bars: 2, chords: [] },
+      { name: "A", bars: 4, chords: [] },
+      { name: "B", bars: 4, chords: [] },
+      { name: "Outro", bars: 2, chords: [] },
+    ],
+    bpm: 92,
+    key: "A",
+    mood: ["cozy"],
+    instruments: ["rhodes", "upright bass"],
+    ambience: ["cafe"],
+    drumPattern: "boom_bap_B",
+    variety: 60,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: false,
+    limiterDrive: 1.1,
+    dither: true,
+    bpmJitterPct: 8,
+  },
+  "New Fantasy": {
+    structure: [
+      { name: "Intro", bars: 4, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "B", bars: 8, chords: [] },
+      { name: "C", bars: 8, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "Outro", bars: 4, chords: [] },
+    ],
+    bpm: 78,
+    key: "E",
+    mood: ["fantasy", "nostalgic", "melancholy"],
+    instruments: ["flute", "piano", "upright bass"],
+    ambience: ["rain"],
+    drumPattern: "swing",
+    variety: 35,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.05,
+    dither: true,
+    bpmJitterPct: 4,
+  },
+  "Sleep": {
+    structure: [
+      { name: "Intro", bars: 4, chords: [] },
+      { name: "A", bars: 12, chords: [] },
+      { name: "B", bars: 12, chords: [] },
+      { name: "Outro", bars: 4, chords: [] },
+    ],
+    bpm: 60,
+    key: "Auto",
+    mood: ["calm", "cozy"],
+    instruments: ["piano", "pads", "upright bass"],
+    ambience: ["rain"],
+    drumPattern: "no_drums",
+    variety: 15,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: false,
+    limiterDrive: 0.98,
+    dither: true,
+    bpmJitterPct: 2,
+  },
+  "Sunset Sketches": {
+    structure: [
+      { name: "Intro", bars: 4, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "B", bars: 8, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "Break", bars: 4, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "Outro", bars: 4, chords: [] },
+    ],
+    bpm: 82,
+    key: "Auto",
+    mood: ["calm", "nostalgic", "cozy"],
+    instruments: ["rhodes", "upright bass", "pads"],
+    ambience: ["rain"],
+    leadInstrument: "synth lead",
+    drumPattern: "boom_bap_A",
+    variety: 38,
+    chordSpanBeats: 4,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.04,
+    dither: true,
+    bpmJitterPct: 4,
+  },
+  "Neon Rain": {
+    structure: [
+      { name: "Intro", bars: 2, chords: [] },
+      { name: "Verse", bars: 16, chords: [] },
+      { name: "Chorus", bars: 8, chords: [] },
+      { name: "Verse", bars: 16, chords: [] },
+      { name: "Chorus", bars: 8, chords: [] },
+      { name: "Outro", bars: 4, chords: [] },
+    ],
+    bpm: 88,
+    key: "Em",
+    mood: ["melancholy", "nostalgic"],
+    instruments: ["electric piano", "upright bass", "airy pads"],
+    ambience: ["rain", "street"],
+    leadInstrument: "saxophone",
+    drumPattern: "half_time",
+    variety: 52,
+    chordSpanBeats: 8,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.12,
+    dither: true,
+    bpmJitterPct: 6,
+  },
+  "Loft Morning": {
+    structure: [
+      { name: "Ambient", bars: 4, chords: [] },
+      { name: "A", bars: 12, chords: [] },
+      { name: "A", bars: 12, chords: [] },
+      { name: "B", bars: 12, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "Outro", bars: 4, chords: [] },
+    ],
+    bpm: 70,
+    key: "G",
+    mood: ["calm", "cozy"],
+    instruments: ["piano", "upright bass", "pads"],
+    ambience: ["cafe"],
+    leadInstrument: "flute",
+    drumPattern: "laidback",
+    variety: 22,
+    chordSpanBeats: 4,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: false,
+    limiterDrive: 1.0,
+    dither: true,
+    bpmJitterPct: 3,
+  },
+  "Late Night Coding": {
+    structure: [
+      { name: "Intro", bars: 2, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "B", bars: 8, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "C", bars: 8, chords: [] },
+      { name: "B", bars: 8, chords: [] },
+      { name: "Outro", bars: 2, chords: [] },
+    ],
+    bpm: 92,
+    key: "D",
+    mood: ["cozy"],
+    instruments: ["clean electric guitar", "upright bass", "pads"],
+    ambience: ["cafe"],
+    leadInstrument: "synth lead",
+    drumPattern: "boom_bap_B",
+    variety: 58,
+    chordSpanBeats: 2,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.08,
+    dither: true,
+    bpmJitterPct: 7,
+  },
+  "Forest Spirits": {
+    structure: [
+      { name: "Intro", bars: 4, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "B", bars: 8, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "Ambient", bars: 4, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "Outro", bars: 4, chords: [] },
+    ],
+    bpm: 76,
+    key: "C",
+    mood: ["fantasy", "nostalgic", "calm"],
+    instruments: ["harp", "piano", "upright bass", "airy pads"],
+    ambience: ["rain", "street"],
+    leadInstrument: "flute",
+    drumPattern: "swing",
+    variety: 34,
+    chordSpanBeats: 8,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.03,
+    dither: true,
+    bpmJitterPct: 5,
+  },
+  "Linear Drift": {
+    structure: [
+      { name: "Intro", bars: 4, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "B", bars: 8, chords: [] },
+      { name: "C", bars: 8, chords: [] },
+      { name: "D", bars: 8, chords: [] },
+      { name: "E", bars: 8, chords: [] },
+      { name: "F", bars: 8, chords: [] },
+      { name: "Outro", bars: 4, chords: [] },
+    ],
+    bpm: 82,
+    key: "Auto",
+    mood: ["calm", "nostalgic"],
+    instruments: ["rhodes", "upright bass", "pads"],
+    ambience: ["rain"],
+    drumPattern: "laidback",
+    variety: 42,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.0,
+    dither: true,
+    bpmJitterPct: 4,
+  },
+  "Odd Odyssey": {
+    structure: [
+      { name: "Intro", bars: 4, chords: [] },
+      { name: "A", bars: 7, chords: [] },
+      { name: "B", bars: 5, chords: [] },
+      { name: "C", bars: 7, chords: [] },
+      { name: "D", bars: 5, chords: [] },
+      { name: "E", bars: 7, chords: [] },
+      { name: "F", bars: 5, chords: [] },
+      { name: "Outro", bars: 4, chords: [] },
+    ],
+    bpm: 78,
+    key: "Auto",
+    mood: ["cozy", "nostalgic"],
+    instruments: ["piano", "upright bass", "pads"],
+    ambience: ["street"],
+    drumPattern: "swing",
+    variety: 45,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.05,
+    dither: true,
+    bpmJitterPct: 5,
+  },
+  "Arcane Clash": {
+    structure: [
+      { name: "Intro", bars: 2, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "B", bars: 8, chords: [] },
+      { name: "C", bars: 8, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "Outro", bars: 2, chords: [] },
+    ],
+    bpm: 110,
+    key: "Dm",
+    mood: ["fantasy", "melancholy"],
+    instruments: ["violin", "cello", "trumpet", "timpani"],
+    ambience: ["forest"],
+    leadInstrument: "violin",
+    drumPattern: "half_time",
+    variety: 70,
+    chordSpanBeats: 4,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.15,
+    dither: true,
+    bpmJitterPct: 3,
+  },
+  "King's Last Stand": {
+    structure: [
+      { name: "Intro", bars: 4, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "B", bars: 8, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "Boss", bars: 8, chords: [] },
+      { name: "Outro", bars: 4, chords: [] },
+    ],
+    bpm: 96,
+    key: "Em",
+    mood: ["fantasy", "melancholy"],
+    instruments: ["pipe organ", "violin", "cello", "trumpet", "timpani", "choir"],
+    ambience: ["forest"],
+    leadInstrument: "trumpet",
+    drumPattern: "half_time",
+    variety: 60,
+    chordSpanBeats: 8,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.2,
+    dither: true,
+    bpmJitterPct: 5,
+  },
+  "Ocean Breeze": {
+    structure: [
+      { name: "Intro", bars: 4, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "B", bars: 8, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "Outro", bars: 4, chords: [] },
+    ],
+    bpm: 72,
+    key: "G",
+    mood: ["calm", "nostalgic", "cozy"],
+    instruments: ["acoustic guitar", "piano", "upright bass", "airy pads"],
+    ambience: ["ocean"],
+    leadInstrument: "flute",
+    drumPattern: "laidback",
+    variety: 30,
+    chordSpanBeats: 4,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.0,
+    dither: true,
+    bpmJitterPct: 3,
+  },
+  "City Lights": {
+    structure: [
+      { name: "Intro", bars: 2, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "B", bars: 8, chords: [] },
+      { name: "C", bars: 8, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "Outro", bars: 2, chords: [] },
+    ],
+    bpm: 92,
+    key: "F#",
+    mood: ["cozy", "nostalgic"],
+    instruments: ["electric piano", "upright bass", "synth lead", "shaker"],
+    ambience: ["street", "rain"],
+    leadInstrument: "saxophone",
+    drumPattern: "half_time",
+    variety: 55,
+    chordSpanBeats: 4,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.1,
+    dither: true,
+    bpmJitterPct: 6,
+  },
+  "Starlit Voyage": {
+    structure: [
+      { name: "Intro", bars: 4, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "B", bars: 8, chords: [] },
+      { name: "Ambient", bars: 4, chords: [] },
+      { name: "A", bars: 8, chords: [] },
+      { name: "Outro", bars: 4, chords: [] },
+    ],
+    bpm: 80,
+    key: "C#",
+    mood: ["calm", "melancholy"],
+    instruments: ["synth lead", "pads", "upright bass", "shaker"],
+    ambience: ["vinyl"],
+    leadInstrument: "synth lead",
+    drumPattern: "half_time",
+    variety: 40,
+    chordSpanBeats: 8,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.05,
+    dither: true,
+    bpmJitterPct: 4,
+  },
+  "Sunset VHS": {
+    structure: [
+      { name: "Intro", bars: 4, chords: ["Am7", "Am7", "Am7", "Am7"] },
+      {
+        name: "A",
+        bars: 8,
+        chords: ["Am7", "Gmaj7", "Fmaj7", "E7", "Am7", "Gmaj7", "Fmaj7", "E7"],
+      },
+      {
+        name: "B",
+        bars: 8,
+        chords: ["Dm7", "G7", "Cmaj7", "Fmaj7", "Bm7b5", "E7", "Am7", "Am7"],
+      },
+      {
+        name: "A",
+        bars: 8,
+        chords: ["Am7", "Gmaj7", "Fmaj7", "E7", "Am7", "Gmaj7", "Fmaj7", "E7"],
+      },
+      { name: "Outro", bars: 4, chords: ["Fmaj7", "E7", "Am7", "Am7"] },
+    ],
+    bpm: 90,
+    key: "Am",
+    mood: ["nostalgic", "dreamy"],
+    instruments: ["wurlitzer", "upright bass", "pads", "synth lead"],
+    ambience: ["vinyl"],
+    ambienceLevel: 0.35,
+    drumPattern: "laidback",
+    variety: 45,
+    chordSpanBeats: 4,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.05,
+    dither: true,
+    bpmJitterPct: 5,
+  },
+  "Neon Palms": {
+    structure: [
+      { name: "Intro", bars: 4, chords: ["F#m7", "F#m7", "F#m7", "F#m7"] },
+      {
+        name: "A",
+        bars: 8,
+        chords: ["F#m7", "Dmaj7", "E7", "C#m7", "F#m7", "Dmaj7", "E7", "C#m7"],
+      },
+      {
+        name: "B",
+        bars: 8,
+        chords: ["Amaj7", "B7", "G#m7", "C#7", "Amaj7", "B7", "G#m7", "C#7"],
+      },
+      {
+        name: "A",
+        bars: 8,
+        chords: ["F#m7", "Dmaj7", "E7", "C#m7", "F#m7", "Dmaj7", "E7", "C#m7"],
+      },
+      { name: "Outro", bars: 4, chords: ["Dmaj7", "E7", "F#m7", "F#m7"] },
+    ],
+    bpm: 92,
+    key: "F#m",
+    mood: ["nostalgic", "dreamy"],
+    instruments: ["wurlitzer", "upright bass", "pads", "synth lead"],
+    ambience: ["vinyl"],
+    ambienceLevel: 0.3,
+    drumPattern: "boom_bap_B",
+    variety: 50,
+    chordSpanBeats: 4,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.05,
+    dither: true,
+    bpmJitterPct: 5,
+  },
+  "Night Swim": {
+    structure: [
+      { name: "Intro", bars: 4, chords: ["Dm7", "Dm7", "Dm7", "Dm7"] },
+      {
+        name: "A",
+        bars: 8,
+        chords: ["Dm7", "Gm7", "C7", "Fmaj7", "Dm7", "Gm7", "C7", "Fmaj7"],
+      },
+      {
+        name: "B",
+        bars: 8,
+        chords: ["Bbmaj7", "A7", "Dm7", "Gm7", "Bbmaj7", "A7", "Dm7", "Gm7"],
+      },
+      {
+        name: "A",
+        bars: 8,
+        chords: ["Dm7", "Gm7", "C7", "Fmaj7", "Dm7", "Gm7", "C7", "Fmaj7"],
+      },
+      { name: "Outro", bars: 4, chords: ["Gm7", "A7", "Dm7", "Dm7"] },
+    ],
+    bpm: 84,
+    key: "Dm",
+    mood: ["nostalgic", "calm"],
+    instruments: ["clean electric guitar", "upright bass", "pads", "synth lead"],
+    ambience: ["vinyl"],
+    ambienceLevel: 0.4,
+    drumPattern: "half_time",
+    variety: 40,
+    chordSpanBeats: 4,
+    hqStereo: true,
+    hqReverb: true,
+    hqSidechain: true,
+    hqChorus: true,
+    limiterDrive: 1.05,
+    dither: true,
+    bpmJitterPct: 5,
+  },
+};
